+++ examples/tictactoe/api/TicTacToeActor.ts
@@ -1,7 +1,8 @@
 import { Schema as S } from "effect"
 import { Actor } from "liminal"
 
-import { TicTacToeClient, Player } from "./TicTacToeClient.ts"
+import { TicTacToeClient } from "./TicTacToeClient.ts"
+import { Player } from "./domain.ts"
 
 export class TicTacToeActor extends Actor.Service<TicTacToeActor>()("examples/TicTacToeActor", {
   client: TicTacToeClient,
