+++ examples/tictactoe/api/errors.ts
@@ -1,0 +1,5 @@
+import { Schema as S } from "effect"
+
+export class OutOfTurnError extends S.TaggedErrorClass<OutOfTurnError>()("OutOfTurnError", {}) {}
+
+export class SlotTakenError extends S.TaggedErrorClass<SlotTakenError>()("SlotTakenError", {}) {}
